--- src/main/java/com/kodilla/ecommercee/OrderDto.java
@@ -1,5 +1,0 @@
-package com.kodilla.ecommercee;
-
-public class OrderDto {
-
-}
