+++ src/main/java/com/kodilla/ecommercee/controller/GroupController.java
@@ -1,6 +1,6 @@
 package com.kodilla.ecommercee.controller;
 
-import com.kodilla.ecommercee.domain.GroupDto;
+import com.kodilla.ecommercee.domain.dto.GroupDto;
 import org.springframework.web.bind.annotation.*;
 
 import javax.websocket.server.PathParam;
@@ -13,25 +13,26 @@ import java.util.List;
 @RequestMapping("/v1/group")
 public class GroupController {
 
-    @RequestMapping(method = RequestMethod.GET)
+    @GetMapping("/groups")
     public List<GroupDto> getGroups() {
         return new ArrayList<>();
     }
 
-
-    @RequestMapping(method = RequestMethod.GET, value = "/{groupId}")
+    @GetMapping("/{groupId}")
     public GroupDto getGroup(@PathParam ("groupId") Long groupId) {
         return new GroupDto();
     }
 
+    @DeleteMapping("/{groupId}")
+    public void deleteGroupById(@PathParam("groupId") Long groupId){
+    }
 
-    @RequestMapping(method = RequestMethod.PUT)
+    @PutMapping("")
     public GroupDto updateGroup(@RequestBody GroupDto groupDto) {
         return new GroupDto();
     }
 
-    @RequestMapping(method = RequestMethod.POST)
+    @PostMapping("")
     public void createGroup(@RequestBody GroupDto groupDto) {
-
     }
 }
