--- src/main/java/com/kodilla/ecommercee/controller/ProductDto.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee.controller;
-
-public class ProductDto {
-}
