--- src/main/java/com/kodilla/ecommercee/domain/GroupDto.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee.domain;
-
-public class GroupDto {
-}
