--- src/main/java/com/kodilla/ecommercee/domain/OrderDto.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee.domain;
-
-public class OrderDto {
-}
