+++ src/main/java/com/kodilla/ecommercee/domain/dto/CartDto.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.domain.dto;
+
+public class CartDto {
+}
