+++ src/main/java/com/kodilla/ecommercee/domain/dto/OrderDto.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.domain.dto;
+
+public class OrderDto {
+}
