+++ src/main/java/com/kodilla/ecommercee/domain/dto/ProductDto.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.domain.dto;
+
+public class ProductDto {
+}
